% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{get_platform_info}
\alias{get_platform_info}
\title{Get Platform Information}
\usage{
get_platform_info()
}
\value{
A list with platform and extension, or NULL if unsupported
}
\description{
Get Platform Information
}
\keyword{internal}
